requirements.txt
pytest~=7.1.1
PyYAML~=6.0
psycopg2-binary~=2.9.3
sshtunnel~=0.4.0
psycopg2~=2.9.3
Pillow~=9.1.0
Faker~=13.3.4
tkcalendar~=1.6.1
pip~=21.2.3
cryptography~=36.0.2
py~=1.11.0
setuptools~=57.4.0